feat(plasma-new-hope): refactor Spinner folder structure
