+++ packages/plasma-new-hope/src/components/Spinner/Spinner.tokens.ts
@@ -1,0 +1,4 @@
+export const tokens = {
+    size: '--plasma-spinner-size',
+    color: '--plasma-spinner-color',
+};
