+++ packages/plasma-new-hope/src/components/Spinner/Spinner.types.ts
@@ -1,0 +1,53 @@
+export type SpinnerInnerProps = {
+    wrapperWidth: string;
+    wrapperHeight: string;
+    svgWidth: string | number;
+    svgHeight: string | number;
+    color: string;
+};
+
+export type SpinnerProps = React.HTMLAttributes<HTMLDivElement> &
+    (
+        | {
+              width?: never;
+              height?: never;
+              /**
+               * Размер спиннера
+               */
+              size?: string;
+              /**
+               * Цвет спиннера
+               */
+              view?: string;
+          }
+        | {
+              /**
+               * Кастомная ширина спиннера
+               */
+              width: number;
+              /**
+               * Кастомная высота спиннера
+               */
+              height: number;
+              size?: never;
+              /**
+               * Цвет спиннера
+               */
+              view?: string;
+          }
+        | {
+              /**
+               * Кастомная ширина спиннера
+               */
+              width: string;
+              /**
+               * Кастомная высота спиннера
+               */
+              height: string;
+              size?: never;
+              /**
+               * Цвет спиннера
+               */
+              view?: string;
+          }
+    );
