+++ packages/plasma-new-hope/src/components/Spinner/index.ts
@@ -1,1 +1,2 @@
 export { spinnerRoot, spinnerConfig } from './Spinner';
+export { tokens as spinnerTokens } from './Spinner.tokens';
