+++ packages/plasma-new-hope/src/components/Spinner/variations/_size/tokens.json
