+++ packages/plasma-new-hope/src/components/Spinner/variations/_view/base.ts
