+++ packages/plasma-new-hope/src/examples/plasma_b2c/components/Spinner/Spinner.config.ts
@@ -1,5 +1,7 @@
 import { css } from '@linaria/core';
 
+import { spinnerTokens } from '../../../../components/Spinner';
+
 export const config = {
     defaults: {
         view: 'accent',
@@ -8,39 +10,39 @@ export const config = {
     variations: {
         size: {
             s: css`
-                --plasma-spinner-size: 16px;
+                ${spinnerTokens.size}: 16px;
             `,
             m: css`
-                --plasma-spinner-size: 24px;
+                ${spinnerTokens.size}: 24px;
             `,
             l: css`
-                --plasma-spinner-size: 36px;
+                ${spinnerTokens.size}: 36px;
             `,
         },
         view: {
             primary: css`
-                --plasma-spinner-color: var(--text-primary);
+                ${spinnerTokens.color}: var(--text-primary);
             `,
             secondary: css`
-                --plasma-spinner-color: var(--text-secondary);
+                ${spinnerTokens.color}: var(--text-secondary);
             `,
             tertiary: css`
-                --plasma-spinner-color: var(--text-tertiary);
+                ${spinnerTokens.color}: var(--text-tertiary);
             `,
             paragraph: css`
-                --plasma-spinner-color: var(--text-paragraph);
+                ${spinnerTokens.color}: var(--text-paragraph);
             `,
             accent: css`
-                --plasma-spinner-color: var(--text-accent);
+                ${spinnerTokens.color}: var(--text-accent);
             `,
             positive: css`
-                --plasma-spinner-color: var(--text-positive);
+                ${spinnerTokens.color}: var(--text-positive);
             `,
             warning: css`
-                --plasma-spinner-color: var(--text-warning);
+                ${spinnerTokens.color}: var(--text-warning);
             `,
             negative: css`
-                --plasma-spinner-color: var(--text-negative);
+                ${spinnerTokens.color}: var(--text-negative);
             `,
         },
     },
